Allow properties to carry over for attributes
When an attribute is on a service, properties with a public getter and setter will be included for processing by the middleware. It does this by copying the existing attribute onto the new attribute generated.

A new attribute is generated to populate constructor fields for dependency injection. The initial attribute on the service wouldn't have the dependencies filled in since it is constructed on the top of the service.

The "Example2Service" utilizes the new feature as an example. The ExampleProperty is filled in on top of the service. Then for the middleware, if you put a break point in ExampleAttribute.ExecuteAsync and call the service, you can see ExampleProperty is still filled in.

Issue: https://github.com/TownSuite/TownSuite.Web.SSV3Adapter/issues/36