Attribute interface to ignore a constructor
Any attribute inheriting the new IIgnoreConstructorAttribute can be used to ensure the constructor isn't used to create the instance in the middleware.

Normally when a service is being constructed the constructor prioritized is the 1 with the most parameters. Now constructors can be explicitly labeled to not be used when constructing the service. If the new attribute is used on the constructor of an attribute on the service, then it will also ignore that constructor when instancing the attribute.

The initial goal with these changes is to allow specifying permission tags in the constructor of the attribute, so that permissions can be defined on the top of services instead of within request objects.

The Example2Service and ExampleAttribute show the use of taking in a string[] exampleProperty parameter into a constructor. Without the ExampleIgnoreConstructorAttribute that inherits the new interface, this is unlikely to work because the constructor would be used for dependency injection (string[] is not injectible).
Issue: https://github.com/TownSuite/TownSuite.Web.SSV3Adapter/issues/36