Don't copy attribute properties with private setters
Attribute properties with private getters or setters will no longer be carried over when the attribute is recreated.

The attribute is recreated for the middleware when calling a service. The attribute would be on top of the service class. The copying occurs to allow the attribute to use dependency injection.

Before this all properties with a getter and setter, even if the getter/setter was private would be copied over to other properties. Now it will only copy if the getter/setter are both public.

Issue: https://github.com/TownSuite/TownSuite.Web.SSV3Adapter/issues/36