Attribute Property carry over example
Example2Service is no longer used to help test adding values into the constructor of attribute on top of the service for use in the middleware. Instead a new ExampleAttributePropertyService was created for it. The value in the request gets changed in the middleware to what ever is in the attributes constructor on top of the service and then is returned in the response.

Issue: https://github.com/TownSuite/TownSuite.Web.SSV3Adapter/issues/36